Updated parseInt to parseFloat

+++ routes/api.js
@@ -35,7 +35,7 @@ router.put('/:id/add/:store/:amount/:date', async (req, res) => {
                 {
                     "uniqid": id,
                     "store": req.params.store,
-                    "amount": parseInt(req.params.amount),
+                    "amount": parseFloat(req.params.amount),
                     "date": req.params.date
                 }
             ]
